p4i integration Integrate 207576 from 4.1 to main: Fixes:    Compiler warning.
[git-p4: depot-paths = "//depot/qt/main/": change = 207577]

+++ src/gui/painting/qblackraster.c
@@ -1712,7 +1712,7 @@ Convert_Glyph( TRaster_Instance*  raster, int  flipped )
         ras.gProfile = NULL;
 
         if ( Decompose_Curve( raster, (unsigned short)start,
-                              ras.outline.contours[i],
+                              (unsigned short)ras.outline.contours[i],
                               flipped ) )
             return FAILURE;
